Compatibility with load balancers : first version
This first verison adds the folowing:

- Reliability of the ADD_ADDR (RFC6824bis) for IPv4
- Support for the NO_JOIN (RFC6824bis) option
- A load-balancing path manager prototype

For the moment, the path manager generates IPs by flipping a specific
high order bit. This will be changed, when a consus is reached on
how we should generate public IP addresses.
The patch adding the reliablity to ADD_ADDR is going to be pushed
upstream, as soon as the IPv6 support is completed.